requirements.txt
opencv-python
numpy
pillow
python-snap7
tqdm
